% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizers.R
\name{lmer_preds}
\alias{lmer_preds}
\title{Format the predicted values from an lmer model.}
\usage{
lmer_preds(frame, x_pred_vars, lmer_mod)
}
\arguments{
\item{frame}{The data.frame used to fit the model.}

\item{x_pred_vars}{A character vector of columns in frame used as
predictors in the model.}

\item{lmer_mod}{An lmer model object to obtain predictions from.}
}
\description{
Format the predicted values from an lmer model.
}
