% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoders.R
\name{recode_age}
\alias{recode_age}
\title{Recode the age column.}
\usage{
recode_age(frame)
}
\description{
Create new columns based off the age column in the frame.
}
\details{
The following columns will be created:
\itemize{
  \item age_sqr. The age variable squared.
}
}

