% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoders.R
\name{recode_edits}
\alias{recode_edits}
\title{Recode the edits column.}
\usage{
recode_edits(frame)
}
\description{
Create new columns based off the edits column in the frame.
}
\details{
The following columns will be created:
\itemize{
  \item edits_sum. The total number of edits to this article so far.
  \item edits_sqr. The edits_sum variable squared.
}
}

