% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoders.R
\name{recode_generations}
\alias{recode_generations}
\title{Recode the generations column.}
\usage{
recode_generations(frame)
}
\description{
Create new columns based off the generations column in the frame.
}
\details{
The following columns will be created:
\itemize{
  \item generations_sum. The total number of generations to this article
        so far.
  \item generations_sum_sqr. The generations_sum variable squared.
}
}

