% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoders.R
\name{recode_quality}
\alias{recode_quality}
\title{Recode the quality column.}
\usage{
recode_quality(frame)
}
\description{
Create new columns based off quality column in frame.
}
\details{
The following columns will be created:

\itemize{
  \item quality_diff. The difference in quality between the current version
        and the previous.
}
}
\examples{
library(wikischolarlib)
data("random1000")
random1000 <- recode_quality(random1000)

}

